+++ atomqueue.h
@@ -38,14 +38,14 @@ typedef struct atom_queue
     ATOM_TCB *  putSuspQ;       /* Queue of threads waiting to send */
     ATOM_TCB *  getSuspQ;       /* Queue of threads waiting to receive */
     uint8_t *   buff_ptr;       /* Pointer to queue data area */
-    uint32_t    unit_size;      /* Size of each message */
-    uint32_t    max_num_msgs;   /* Max number of storable messages */
-    uint32_t    insert_index;   /* Next byte index to insert into */
-    uint32_t    remove_index;   /* Next byte index to remove from */
+    size_t      unit_size;      /* Size of each message */
+    size_t      max_num_msgs;   /* Max number of storable messages */
+    size_t      insert_index;   /* Next byte index to insert into */
+    size_t      remove_index;   /* Next byte index to remove from */
     uint32_t    num_msgs_stored;/* Number of messages stored */
 } ATOM_QUEUE;
 
-extern uint8_t atomQueueCreate (ATOM_QUEUE *qptr, void *buff_ptr, uint32_t unit_size, uint32_t max_num_msgs);
+extern uint8_t atomQueueCreate (ATOM_QUEUE *qptr, void *buff_ptr, size_t unit_size, size_t max_num_msgs);
 extern uint8_t atomQueueDelete (ATOM_QUEUE *qptr);
 extern uint8_t atomQueueGet (ATOM_QUEUE *qptr, int32_t timeout, void *msgptr);
 extern uint8_t atomQueuePut (ATOM_QUEUE *qptr, int32_t timeout, const void *msgptr);
